refactor(interceptor): improve implementation
