--- packages/zimic-interceptor/src/http/interceptorWorker/errors/UnregisteredBrowserServiceWorkerError.ts
@@ -1,1 +1,0 @@
-export { default } from '@/interceptor/errors/UnregisteredBrowserServiceWorkerError';
