--- packages/zimic-interceptor/src/http/interceptorWorker/types/msw.ts
@@ -1,2 +1,0 @@
-export type { BrowserMSWWorker, MSWWorker, NodeMSWWorker } from '@/interceptor/types/msw';
-export type { HttpHandler as MSWHttpHandler } from 'msw';
